+++ ProyectoDisquera/src/main/java/com/unicundi/BD/DAOCompraDisco.java
@@ -5,7 +5,6 @@
  */
 package com.unicundi.BD;
 
-import com.unicundi.utilitarios.UCompraCancion;
 import com.unicundi.utilitarios.UCompraDisco;
 import com.unicundi.utilitarios.UUsuario;
 import java.io.Serializable;
@@ -54,7 +53,7 @@ public class DAOCompraDisco implements Serializable {
                 PreparedStatement stmt = conexion.prepareStatement(query);
                 ResultSet resultado = stmt.executeQuery();
                 while (resultado.next()) {
-                    float valorCompra = (float) resultado.getDouble("valor_compra");
+                    int valorCompra = resultado.getInt("valor_compra");
                     Date fechaCompra = resultado.getDate("fecha_compra");
                     String genero = resultado.getString("genero");
                     String nombreCompletoArtista = resultado.getString("nombre_artista") + " " + resultado.getString("apellido_artista");
